+++ tests/FractalInstanceTest.php
@@ -23,9 +23,9 @@ class FractalInstanceTest extends TestCase
     /** @test */
     public function it_returns_a_configured_instance_when_resolving_fractal_using_the_container()
     {
-        $this->app->forgetInstance('laravel-fractal');
+        $this->app->forgetInstance('fractal');
 
-        app('config')->set('laravel-fractal.fractal_class', FractalExtensionClass::class);
+        app('config')->set('fractal.fractal_class', FractalExtensionClass::class);
 
         $this->assertInstanceOf(FractalExtensionClass::class, $this->app->make(Fractal::class));
     }