+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to `laravel-fractal` will be documented in this file
 
+
+## 5.0.0 - 2017-XX-XX
+
+- Laravel 5.5 support, dropped support for all older versions
+- renamed config file from `laravel-fractal` to `fractal`
+- added auto-includes
+
+## 4.5.0 - 2017-08-20
+- add `default_paginator` to config file
+
+
 ## 4.4.0 - 2017-08-20
 - add `fractal_class` to config file